fix: make karma experimental default
