+++ src/test/index.js
@@ -2,31 +2,15 @@
 
 const TASKS = [{
   title: 'Test Node.js',
-  task: (opts) => {
-    const node = require('./node')
-    return node(opts)
-  },
+  task: (opts) => require('./node')(opts),
   enabled: (ctx) => ctx.target.includes('node')
 }, {
   title: 'Test Browser',
-  task: (opts) => {
-    if (opts.eek) {
-      return require('./experimental-browser')(opts)
-    }
-    const browser = require('./browser')
-    return browser.default(opts)
-  },
+  task: (opts) => require('./browser')(opts),
   enabled: (ctx) => ctx.target.includes('browser')
 }, {
   title: 'Test Webworker',
-  task: (opts) => {
-    if (opts.eek) {
-      opts.webworker = true
-      return require('./experimental-browser')(opts)
-    }
-    const browser = require('./browser')
-    return browser.webworker(opts)
-  },
+  task: (opts) => require('./browser')(Object.assign(opts, { webworker: true })),
   enabled: (ctx) => ctx.target.includes('webworker')
 }]
